+++ tests/unit/store/utility.spec.ts
@@ -19,9 +19,6 @@ import {
   isAccentPhrasesTextDifferent,
   buildAudioFileNameFromRawData,
   getToolbarButtonName,
-  createKanaRegex,
-  convertHiraToKana,
-  convertLongVowel,
   getBaseName,
   isOnCommandOrCtrlKeyDown,
   filterCharacterInfosByStyleType,
@@ -282,31 +279,6 @@ test("getToolbarButtonName", () => {
   expect(getToolbarButtonName("存在しないタグ" as ToolbarButtonTagType)).toBe(
     undefined,
   );
-});
-
-describe("createKanaRegex", () => {
-  it("includeSeparationがtrueの場合、読点とクエスチョンも含む", () => {
-    const regex = createKanaRegex(true);
-    expect(regex.test("あいうえお、")).toBe(true);
-    expect(regex.test("かきくけこ？")).toBe(true);
-  });
-
-  it("includeSeparationがfalseの場合、読点とクエスチョンを含まない", () => {
-    const regex = createKanaRegex(false);
-    expect(regex.test("あいうえお、")).toBe(false);
-    expect(regex.test("かきくけこ？")).toBe(false);
-  });
-});
-
-test("convertHiraToKana", () => {
-  expect(convertHiraToKana("あいうえお")).toBe("アイウエオ");
-  expect(convertHiraToKana("がぱをんー")).toBe("ガパヲンー");
-});
-
-test("convertLongVowel", () => {
-  expect(convertLongVowel("アー")).toBe("アア");
-  expect(convertLongVowel("ガー")).toBe("ガア");
-  expect(convertLongVowel("ンー")).toBe("ンン");
 });
 
 test("getBaseName", () => {
